Ctrl-C Cancels Current Line when --sigint-ignore
This patch changes the behavior of Ctrl-C when the --sigint-ignore
option is enabled. Instead of Ctrl-C simply being ignored completely, it
now invokes the function 'com_clear(&glob_buffer, 0)'. This is the same
command assigned to '\c'. In other words, whenever SIGINT is handled,
instead of being ignored, the current prompt/buffer is completely reset
to a new line. This is the same behavior as exists in the PostgreSQL
CLI.

This was the patch I mentioned in the dba.stackexchange.com question at
http://dba.stackexchange.com/a/99290/64489. I've tested the patch on
OS X, having built just the mysqlclient target, and it works as
advertised.

Hope this helps!

--
-a